Implemented enum file
